+++ CHANGELOG.md
@@ -16,9 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fixed inversed inheritance at unnamed aggregate:
     - Contained unnamed aggregate is ignored.
     - Unnamed aggregate no longer inherits from enclosing unnamed aggregate.
-
 - Fixed static route with sole virtual interface as next hop.
   Use IP of virtual interface instead of physical interface.
+- No longer accidently split NAT domain at unmanaged router
+  with attribute 'bind_nat'.
 
 ## [2025-04-17-1102]